+++ src/main/java/frc/robot/subsystems/TelescopeClimb.java
@@ -14,7 +14,8 @@ import com.ctre.phoenix.motorcontrol.can.TalonSRX;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
 public class TelescopeClimb extends SubsystemBase {
- //Motors
+
+  // Motors
   private TalonSRX teleClimbMotor;
 
   public TelescopeClimb() {
@@ -28,10 +29,12 @@ public class TelescopeClimb extends SubsystemBase {
   public void periodic() {
     // This method will be called once per scheduler run
   }
+
   public void setClimbSpeed(double speed) {
-    teleClimbMotor.set(ControlMode.PercentOutput , speed);
+    teleClimbMotor.set(ControlMode.PercentOutput, speed);
   }
-  public int getTelePosition() {
+
+  public double getTelePosition() {
     return teleClimbMotor.getSelectedSensorPosition();
   }
 }
